requirements.txt
flask==1.1.1
flask-sqlalchemy==2.4.1
psycopg2==2.8.4
Flask-JWT==0.3.2
python-consul
python-logstash-async
prometheus_flask_exporter
